+++ src/helpers/links/index.ts
@@ -2,3 +2,4 @@ export * from './getBaseUrl';
 export * from './openLink';
 export * from './getLinkInfo';
 export * from './isUrl';
+export * from './LinkHandler';