+++ src/state/app/actions/index.ts
@@ -3,3 +3,4 @@ export * from './setHomePress';
 export * from './toast';
 export * from './setDrawerOpen';
 export * from './setUnread';
+export * from './setAppLoading';