+++ src/state/app/actions/setAppLoading.ts
@@ -1,0 +1,7 @@
+import { useAppStore } from '@src/state';
+
+export const setAppLoading = (v: boolean): void => {
+  useAppStore.setState((state) => {
+    state.isLoading = v;
+  });
+};